move: 이벤트 타이머 shared 폴더로 이동
